Fix S4 validation bugs and parallel processing issue
Critical bug fixes introduced by previous S4 validation enhancements:

1. ConvergenceDiagnostics validation (R/class-definitions.R:259-265):
   - Fixed: n_replications counts how many times the best solution
     was replicated, NOT the total number of starts
   - Changed validation from equality check to upper bound check
   - Now correctly allows n_replications <= nrow(loglik_table)

2. R3StepResults validation (R/class-definitions.R:195-202):
   - Fixed: class_means matrix has rows=variables, cols=classes
   - Changed validation from ncol to nrow check
   - Now correctly validates length(distal_vars) == nrow(class_means)

3. Parallel processing with log_sum_exp (R/gmm-survey.R:363-366):
   - Fixed: log_sum_exp function not available in parallel workers
   - Added "log_sum_exp" to clusterExport list
   - Resolves "object 'log_sum_exp' not found" error when cores > 1

4. Test expectation fix (tests/testthat/test-convergence.R:85):
   - Fixed: iterations is integer type, not double
   - Changed expect_type from "double" to "integer"

All test failures from devtools::test() should now be resolved.
These were validation logic errors, not issues with the original code.

+++ tests/testthat/test-convergence.R
@@ -82,7 +82,7 @@ test_that("Convergence info is properly stored in object", {
   expect_true("converged" %in% names(fit@convergence_info))
   expect_true("iterations" %in% names(fit@convergence_info))
   expect_type(fit@convergence_info$converged, "logical")
-  expect_type(fit@convergence_info$iterations, "double")
+  expect_type(fit@convergence_info$iterations, "integer")
 })
 
 test_that("Multiple starts with small n converge to similar solutions", {
